DEADSGOLD/requirements.txt
solders
solana
spl-token
cryptography
